Applied patch from Derek Kozel <derek.kozel@gmail.com>:
During the linux-gpib-kernel make install process the kernel 
modules_install command is run bit the DESTDIR argument is not passed 
in meaning that it will always try to install to the system directory 
rather than a custom directory. Adding INSTALL_MOD_PATH="$(DESTDIR)" to 
the command correctly passes the information.


git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1767 b61deed1-901b-0410-a6dd-e94ecbdaea89