Removed obsolete hotplug usermaps.
Fixed problems with usb udev rules not working if the driver
module was not already loaded when the hardware was plugged
in.

Made usb drivers respond to the sysfs device path of their
usb interface rather than the parent usb device.  This is helpful
since the driver binds to the interfaces rather than the device.


git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1781 b61deed1-901b-0410-a6dd-e94ecbdaea89